refactor(hooks): fix correctness issues, add tests, split log-lightweight into modules
- Fix lint error (unused variable), console.error→log(), stale comment
- Standardize observe/main.js to use parseStdinJson + setSessionIdFromInput
- Export checkConsoleLogs from quality-check for testability
- Add 26 new tests for observe, pre-compact, and quality-check hooks
- Split log-lightweight.py (896 lines) into 7-module package (<400 lines each)
- Translate Traditional Chinese strings to English in log schema/summaries
- Extract _make_initial_state factory to eliminate duplicated state dict
- Remove TOCTOU exists() checks in favor of direct read with error handling
- Document hasDevDependency alias, run-hook.cmd naming, SessionStart matchers
